Organize code in ClimaAnalysis extension
